requirements.txt
flask
sqlalchemy
mysqlclient
python-dotenv